+++ changelog.d/287.added.md
@@ -1,1 +1,1 @@
-Added a public `policybench_transfer_2025` dataset constructor that maps PolicyBench's public 1,000-household US benchmark sample into a synthetic `UKSingleYearDataset`, preserving source weights and documenting the result as a UK transfer dataset rather than a representative UK baseline.
+Added a public `enhanced_cps_2025` dataset constructor that maps PolicyBench's public 1,000-household CPS-derived sample into a `UKSingleYearDataset` and recalibrates household weights against the UK national/region/country target registry. Backward-compatible `policybench_transfer` aliases remain available.
