+++ policyengine_uk_data/datasets/__init__.py
@@ -1,3 +1,9 @@
+from .enhanced_cps import (
+    ENHANCED_CPS_FILE,
+    ENHANCED_CPS_SOURCE_FILE,
+    create_enhanced_cps,
+    save_enhanced_cps,
+)
 from .frs import create_frs
 from .policybench_transfer import (
     POLICYBENCH_TRANSFER_SOURCE_FILE,
@@ -7,9 +13,13 @@ from .policybench_transfer import (
 from .spi import create_spi
 
 __all__ = [
+    "ENHANCED_CPS_FILE",
+    "ENHANCED_CPS_SOURCE_FILE",
+    "create_enhanced_cps",
     "POLICYBENCH_TRANSFER_SOURCE_FILE",
     "create_frs",
     "create_policybench_transfer",
     "create_spi",
+    "save_enhanced_cps",
     "save_policybench_transfer",
 ]
